Change get_valid_score function to intergers

+++ prac_02/score_menu.py
@@ -31,10 +31,10 @@ def main():
 
 
 def get_valid_score():
-    score = float(input("Enter score: "))
+    score = int(input("Enter score: "))
     while score < 0 or score > 100:
         print("Invalid score")
-        score = float(input("Enter score: "))
+        score = int(input("Enter score: "))
     return score
